updates progress-tracker and adds test.
There was a bug in my code regarding time interpretation.
This commit adresses it so that this feature does not fail with old postcast entries with a different time format.
Also, as this issue bug could be easily perceived if we had tests, this commit add some basic tests for the utilitary methods used.

fixes #57