Fix the length attribute of enclosure tag - Podcast feed It also adds the audio_file_length to the source files of posts
The length attribute is the size in bytes of the audio file.

You can get this value using cURL. Example:

    curl -I -L http://zofe.com.br/episodios/zofe-0-apresentacao

The "Content-Length" header is what we want. If the URL follow
redirects (which is the case of SoundCloud), you may need to take
the last value for the audio file length (which is the bigger number).